% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addl_expr.R
\docType{data}
\name{addl_expr}
\alias{addl_expr}
\title{Additional expressions for \code{drug_list}}
\format{
A data frame with the following variables:
\describe{
  \item{expr}{A character vector, additional optional expressions for the \code{drug_list} argument.}
  \item{type}{A character vector, what category the expression belongs to (e.g., symptom, lab name, medication abbreviation, or drug class).}
}
}
\usage{
addl_expr
}
\description{
A dictionary with additional expressions that can be used to supplement the \code{drug_list}
argument of \code{\link{medExtractR}} and \code{\link{medExtractR_tapering}}.
}
\examples{
data(addl_expr)
}
\keyword{datasets}
