% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosechange_vals.R
\docType{data}
\name{dosechange_vals}
\alias{dosechange_vals}
\title{Keywords Specifying Dose Change}
\format{
A data frame with dose change expressions (exact and/or regular expressions).
\describe{
  \item{expr}{A character vector, expressions to consider as dose change.}
}
}
\usage{
dosechange_vals
}
\description{
A dictionary of words indicating a dose change, meaning that the associated
drug regimen may not be current. This includes phrases such as increase,
reduce, or switch. In the following example of clinical text, the word
\sQuote{increase} represents a dose change keyword: \dQuote{Increase prograf to 5mg bid.}
}
\examples{
data(dosechange_vals)
}
\keyword{datasets}
