% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseschange_vals.R
\docType{data}
\name{doseschange_vals}
\alias{doseschange_vals}
\title{Keywords Specifying Dose Change}
\format{A vector with 15 character strings.}
\usage{
doseschange_vals
}
\description{
Vector of keywords indicating a dose change.
}
\examples{
data(doseschange_vals)
}
\keyword{datasets}
