% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseschedule_vals.R
\docType{data}
\name{doseschedule_vals}
\alias{doseschedule_vals}
\title{Keywords Specifying Dose Schedule}
\format{
A vector with XX observations on the following variables.
\describe{
  \item{expr}{A character vector, expressions to consider as dose schedule}
}
}
\usage{
doseschedule_vals
}
\description{
A dictionary with words for indicating a tapering dosing schedule
}
\examples{
data(doseschedule_vals)
}
\keyword{datasets}
