% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_vals.R
\docType{data}
\name{duration_vals}
\alias{duration_vals}
\title{Keywords Specifying Duration}
\format{
A data frame with XX observations on the following variables.
\describe{
  \item{expr}{A character vector, expressions to consider as duration}
  \item{value}{A numeric vector, numeric value of duration}?
}
}
\usage{
duration_vals
}
\description{
A dictionary mapping duration expressions to numeric values representing
the corresponding number of doses per day. The form of each duration is
given as a regular expression.
}
\examples{
data(duration_vals)
}
\keyword{datasets}
