% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{entity_metadata}
\alias{entity_metadata}
\title{Grab Entity Metadata}
\usage{
entity_metadata(phrase, p_offset, df)
}
\arguments{
\item{phrase}{Text of interest.}

\item{p_offset}{Position of text within greater document.}

\item{df}{data.frame with position and length of found entities.}
}
\value{
data.frame with found entities and accompanying metadata.
}
\description{
This function standardizes various forms of "tablet" with the string "tabs".
}
\keyword{internal}
