% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_generic.R
\name{extract_generic}
\alias{extract_generic}
\title{Extract Generic Attributes From Phrase}
\usage{
extract_generic(phrase, dict)
}
\arguments{
\item{phrase}{Text to search.}

\item{dict}{data.frame, the first column should contain expressions to find.}
}
\value{
A numeric matrix with position and expression length.
}
\description{
This function searches a phrase for the position and length
of expressions specified in a dictionary.
}
