% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lastdose.R
\name{extract_lastdose}
\alias{extract_lastdose}
\title{Function that extracts time of last dose}
\usage{
extract_lastdose(phrase, p_start, p_stop, p_offset, d_start, d_stop,
  time_exp = "default")
}
\arguments{
\item{phrase}{Text to search}

\item{p_start}{Start position of phrase in larger text.}

\item{p_stop}{Stop position of phrase in larger text.}

\item{p_offset}{Offset in characters used for calibrating last dose time
positions (used when part of phrase was truncated/removed)}

\item{d_start}{Start position of drug name in larger text.}

\item{d_stop}{Start position of drug name in larger text.}

\item{time_exp}{Vector of regular expressions to identify time expressions}
}
\value{
data.frame with last dose time entity information
}
\description{
Function that extracts time of last dose
}
\examples{
extract_lastdose("aspirin last night at 10pm", 25, 50, 0, 25, 31)
}
