% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_type}
\alias{extract_type}
\title{Internal handler for generic entity extraction}
\usage{
extract_type(phrase, type, fun = NULL, ...)
}
\description{
Internal handler for generic entity extraction
}
\keyword{internal}
