% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_vals.R
\docType{data}
\name{freq_vals}
\alias{freq_vals}
\title{Keywords Specifying Frequency}
\format{
A data frame with 77 observations on the following variables.
\describe{
  \item{expr}{A character vector, expressions to consider as frequency}
  \item{value}{A numeric vector, numeric value of frequency}
}
}
\usage{
freq_vals
}
\description{
A dictionary mapping frequency expressions to numeric values representing 
the corresponding number of doses per day. The form of each frequency is
given as a regular expression.
}
\examples{
data(freq_vals)
}
\keyword{datasets}
