% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intaketime_vals.R
\docType{data}
\name{intaketime_vals}
\alias{intaketime_vals}
\title{Keywords Specifying Intake Time}
\format{A data frame with 23 observations on the following variables.
\describe{
  \item{expr}{A character vector, expressions to consider as intake time}
  \item{value}{A numeric vector, numeric value of intake time}
}}
\usage{
intaketime_vals
}
\description{
A dictionary mapping intake time expressions to numeric values representing
the corresponding number of doses. The form of each intake time is given as a
regular expression.
}
\examples{
data(intaketime_vals)
}
\keyword{datasets}
