% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{internal_find_numbers}
\alias{internal_find_numbers}
\title{Preprocess Phrase for Numbers}
\usage{
internal_find_numbers(phrase, num_patt, num_then_ds, num_as_str)
}
\description{
Preprocess Phrase for Numbers
}
\keyword{internal}
