% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medExtractR.R
\name{medExtractR}
\alias{medExtractR}
\title{Extract Medication Entities From Clinical Note}
\usage{
medExtractR(note, drug_names, drug_list, window_length, max_dist, unit,
  strength_sep = NULL, flag_window = 30, doseschange_dict = NULL,
  ...)
}
\arguments{
\item{note}{Text to search.}

\item{drug_names}{Vector of drug names to locate.}

\item{drug_list}{Vector of known drugs that may end expression window.}

\item{window_length}{Length (number of characters) of window after drug
in which to look.}

\item{max_dist}{Numeric - edit distance to use when searching for drug_names.}

\item{unit}{Strength unit to look for (e.g. "mg").}

\item{strength_sep}{Delimiter for contiguous medication strengths.}

\item{flag_window}{How far around drug (number of characters) to look for
dosechange keyword - default fixed to 30.}

\item{doseschange_dict}{List of keywords used to determine if a dose change occurs.}

\item{\dots}{Parameter settings that will be used in extracting
frequency and intake time. Potentially useful parameters include \sQuote{freq_dict}
and \sQuote{intaketime_dict} to specify frequency or intake time keywords, as well
as \sQuote{freq_fun} and \sQuote{intaketime_fun} for user-specifed extraction functions.}
}
\value{
data.frame with entity information
}
\description{
This function extracts dose attributes.
}
