% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preposition_vals.R
\docType{data}
\name{preposition_vals}
\alias{preposition_vals}
\title{Keywords Specifying Preposition}
\format{
A vector with XX observations on the following variables.
\describe{
  \item{expr}{A character vector, expressions to consider as preposition}
}
}
\usage{
preposition_vals
}
\description{
A dictionary with words for prepositions.
}
\examples{
data(preposition_vals)
}
\keyword{datasets}
