% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_tab}
\alias{replace_tab}
\title{String Replace Tablet With Tabs}
\usage{
replace_tab(phrase)
}
\arguments{
\item{phrase}{Text to search.}
}
\value{
Updated phrase.
}
\description{
This function standardizes various forms of "tablet" with the string "tabs".
}
\keyword{internal}
