% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_vals.R
\docType{data}
\name{route_vals}
\alias{route_vals}
\title{Keywords Specifying Route}
\format{
A data frame with the following variables
\describe{
  \item{expr}{A character vector, expressions to consider as route}
  \item{value}{A character vector, standardized route expression}
}
}
\usage{
route_vals
}
\description{
A dictionary mapping route expressions to standardized expressions
The form of each duration is given as a regular expression.
}
\examples{
data(route_vals)
}
\keyword{datasets}
