% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_druglist.R
\docType{data}
\name{rxnorm_druglist}
\alias{rxnorm_druglist}
\title{List of Medications}
\format{A vector with 59,335 character strings.}
\usage{
rxnorm_druglist
}
\description{
Vector of medication names, derived from RxNorm.
RxNorm is provided by the U.S. National Library of Medicine.
}
\details{
THIS NEEDS TO BE INVESTIGATED!
\href{https://uts.nlm.nih.gov/license.html}{UMLS License}
}
\examples{
data(rxnorm_druglist)
}
\keyword{datasets}
