% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_counts.R
\name{string_counts}
\alias{string_counts}
\title{Counts Strings in Text}
\usage{
string_counts(strings, search_data, ignore.case = TRUE)
}
\arguments{
\item{strings}{character vector; value(s) to find}

\item{search_data}{character vector; phrase(s) where values may exist}

\item{ignore.case}{logical; indicates if spelling case matters, defaulting to \sQuote{TRUE}}
}
\value{
list with two elements; \sQuote{cntByTotal} contains total occurrences
and \sQuote{cntByData} contains occurrences for each element in \sQuote{search_data}
}
\description{
This function counts occurrences of text within one or more phrases.
}
\examples{
note1 <- "I am the very model of a modern major general
I've information vegetable, animal, and mineral
I know the kings of England, and I quote the fights historical
From marathon to Waterloo in order categorical;
I'm very well acquainted, too, with matters mathematical,
I understand equations both the simple and quadratical
About binomial theorem I'm teeming with a lot o' news,
With many cheerful facts about the square of the hypotenuse"
note2 <- "The quick brown fox jumps over the lazy dog"
string_counts(c('I','the','couth'), c(note1, note2))
}
