% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{strings-internal}
\alias{strings-internal}
\alias{stringPlaceExact}
\alias{stringPlaceFuzzy}
\alias{stringPlace}
\alias{checkFuzzy}
\title{Internal String Functions}
\usage{
stringPlaceExact(needle, haystack, ignore.case = TRUE)
}
\description{
code{stringPlaceExact}: find location of string, using exact match
}
\details{
code{stringPlaceFuzzy}: find location of string, using partial match

code{stringPlace}: find location of string, generic

code{checkFuzzy}: check for "close" string matches
}
\keyword{internal}
