% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timekeyword_vals.R
\docType{data}
\name{timekeyword_vals}
\alias{timekeyword_vals}
\title{Keywords Specifying Time Keyword}
\format{
A vector with XX observations on the following variables.
\describe{
  \item{expr}{A character vector, expressions to consider as time keyword}
}
}
\usage{
timekeyword_vals
}
\description{
A dictionary with words for indicating a time keyword, i.e. whether the dose is a past, current, or future dose
}
\examples{
data(timekeyword_vals)
}
\keyword{datasets}
