% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_vals.R
\docType{data}
\name{transition_vals}
\alias{transition_vals}
\title{Keywords/Symbols Specifying Transition}
\format{
A vector with XX observations on the following variables.
\describe{
  \item{expr}{A character vector, expressions to consider as transition}
}
}
\usage{
transition_vals
}
\description{
A dictionary with words and symbols used for transitions between segments of a tapering schedule.
}
\examples{
data(transition_vals)
}
\keyword{datasets}
