% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseschedule_vals.R
\docType{data}
\name{doseschedule_vals}
\alias{doseschedule_vals}
\title{Keywords Specifying Dose Schedule}
\format{
A data frame with dose schedule expressions (exact and/or regular expressions).
\describe{
  \item{expr}{A character vector, expressions to consider as dose schedule.}
}
}
\usage{
doseschedule_vals
}
\description{
A dictionary with words for indicating a tapering dosing schedule. These can explicitly
refer to such a schedule with phrases like "tapering" or "wean". It also includes words indicating
an alternating dose schedule (e.g., "alternate", "alt.", "even days", or "odd days") as well as
stopping keywords indicating the patient is going completely off the medication (e.g., "done", "gone",
"stop", "discontinue").
}
\examples{
data(doseschedule_vals)
}
\keyword{datasets}
