% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_vals.R
\docType{data}
\name{duration_vals}
\alias{duration_vals}
\title{Keywords Specifying Duration}
\format{
A data frame with duration expressions (exact and/or regular expressions).
\describe{
  \item{expr}{A character vector, expressions to consider as duration.}
}
}
\usage{
duration_vals
}
\description{
A dictionary with phrases indicating how long the patient should take a particular
dose of the drug. Examples of duration expressions include "2 weeks", "14 days",
"another 3 days", "through mid-April", or a specific date. The form of each duration is
given as a regular expression.
}
\examples{
data(duration_vals)
}
\keyword{datasets}
