% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{entity_metadata}
\alias{entity_metadata}
\title{Grab Entity Metadata}
\usage{
entity_metadata(phrase, p_offset, df)
}
\arguments{
\item{phrase}{Text of interest.}

\item{p_offset}{Start position of text within greater document.}

\item{df}{data.frame with position and length of found medication entities}
}
\value{
data.frame with found medication entities and accompanying metadata.
}
\description{
This function builds the metadata based on position.
}
\keyword{internal}
