% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_entities.R
\name{extract_entities}
\alias{extract_entities}
\title{Extract Entities From Phrase}
\usage{
extract_entities(phrase, p_start, p_stop, unit, freq_fun = NULL,
  intaketime_fun = NULL, strength_sep = NULL, ...)
}
\arguments{
\item{phrase}{Text to search.}

\item{p_start}{Start position of phrase within original text.}

\item{p_stop}{End position of phrase within original text.}

\item{unit}{Unit of measurement for medication strength.}

\item{freq_fun}{Function used to extract frequency.}

\item{intaketime_fun}{Function used to extract intaketime.}

\item{strength_sep}{Delimiter for contiguous medication strengths.}

\item{\dots}{Parameter settings that will be used in extracting
frequency and intake time.}
}
\value{
data.frame with entity information
}
\description{
This function searches a phrase for expressions of interest.
}
