% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intaketime_vals.R
\docType{data}
\name{intaketime_vals}
\alias{intaketime_vals}
\title{Keywords Specifying Intake Time}
\format{
A data frame with intake time expressions (exact and/or regular expressions).
\describe{
  \item{expr}{A character vector, expressions to consider as intake time.}
}
}
\usage{
intaketime_vals
}
\description{
A dictionary with intake time expressions representing the approximate time of
day when a dose should be taken. Example expressions include "in the morning",
"with lunch", "at bedtime", and "qpm". The form of each intake time is
given as a regular expression.
}
\examples{
data(intaketime_vals)
}
\keyword{datasets}
