% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_tab}
\alias{replace_tab}
\title{String Replace Tablet With Tabs}
\usage{
replace_tab(phrase)
}
\arguments{
\item{phrase}{Text to search.}
}
\value{
Updated phrase.
}
\description{
This function standardizes various forms of \dQuote{tablet} or
\dQuote{capsule} with the string \dQuote{tabs}.
}
\details{
Some entities use a keyword such as tablet, capsule, or pill to
identify the entities, e.g. dose amount. This standardizes possible
expressions of similar keywords.
}
\keyword{internal}
