% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timekeyword_vals.R
\docType{data}
\name{timekeyword_vals}
\alias{timekeyword_vals}
\title{Keywords Specifying Time Keyword}
\format{
A data frame with time keyword expressions (exact and/or regular expressions).
\describe{
  \item{expr}{A character vector, expressions to consider as time keyword.}
}
}
\usage{
timekeyword_vals
}
\description{
A dictionary with time keyword expressions representing whether the dosing regimen is
past, current, or future. Example expressions include "currently", "remain","not taking",
"yesterday", and "past".
}
\examples{
data(timekeyword_vals)
}
\keyword{datasets}
