% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_vals.R
\docType{data}
\name{transition_vals}
\alias{transition_vals}
\title{Keywords/Symbols Specifying Transition}
\format{
A data frame with transition expressions (exact and/or regular expressions).
\describe{
  \item{expr}{A character vector, expressions to consider as transitions.}
}
}
\usage{
transition_vals
}
\description{
A dictionary with transition symbols and expressions representing a break between consecutive
doses within a tapering regimen. This dictionary includes the expressions "then" and
"followed by", as well as the punctuation ",(?!\\\\s?then)" or ";(?!\\\\s?then)" (i.e., a
comma or semicolon not followed by the word "then").
}
\examples{
data(transition_vals)
}
\keyword{datasets}
